sis-api/requirements-dev.txt
-r requirements.txt
coverage==4.5.4
django-jenkins==0.110.0
freezegun==1.0.0
ipython==7.16.3
jedi==0.17.2
pep8==1.7.1
pylint==1.9.5
  isort==4.3.21
  astroid==1.6.6
pycodestyle==2.6.0
factory-boy==2.11.1
